requirements.txt
pycryptodome==3.18.0
python-dotenv==1.0.0
web3==6.3.0
websockets==10.4
